+++ lib/Model/PreviewFile.php
@@ -67,7 +67,7 @@ class PreviewFile implements ModelInterface, ArrayAccess
       * @var string[]
       */
     protected static $swaggerFormats = [
-        'id' => 'int32',
+        'id' => 'int64',
 'type' => null,
 'attributes' => null    ];
